Enhancement: Extract BetweenMinimumAndMaximumGreaterThanZeroCountResolutionStrategy
