+++ test/Unit/FixtureFactory/UsingWithOptionalStrategyTest.php
@@ -33,6 +33,7 @@ use PHPUnit\Framework;
 #[Framework\Attributes\UsesClass(FieldDefinition\References::class)]
 #[Framework\Attributes\UsesClass(FieldDefinition\Sequence::class)]
 #[Framework\Attributes\UsesClass(FieldDefinition\Value::class)]
+#[Framework\Attributes\UsesClass(FieldResolution\BetweenMinimumAndMaximumGreaterThanZeroCountResolutionStrategy::class)]
 #[Framework\Attributes\UsesClass(FieldResolution\DefaultStrategy::class)]
 #[Framework\Attributes\UsesClass(FieldResolution\WithOptionalStrategy::class)]
 final class UsingWithOptionalStrategyTest extends Test\Unit\AbstractTestCase
